requirements.txt
flask
markdown2
selenium
bardapi
python-dotenv
google-generativeai
chromedriver-autoinstaller
gunicorn